File upload and send email Hide show changes and Custom layout changes
