+++ libs/shared/src/lib/components/email/components/email-attachment/email-attachment.component.html
@@ -2,22 +2,17 @@
 <div class="bg-[#F5FAFF] mb-4 p-5">
   <label class="block text-base font-medium leading-6 text-gray-900">{{
     'common.notifications.email.attachment.title' | translate
-  }}</label>
-  <div
-    class="flex items-center justify-between gap-5 mb-5 border border-gray-300 rounded-md p-4"
-  >
+    }}</label>
+  <div class="flex items-center justify-between gap-5 mb-5 border border-gray-300 rounded-md p-4">
     <!-- File Upload Button -->
     <div class="w-full">
-      <label
-        for="fileUpload"
-        class="block text-sm font-normal leading-6 text-gray-900"
-      >
+      <label for="fileUpload" class="block text-sm font-normal leading-6 text-gray-900">
         {{ 'common.notifications.email.attachment.uploadFile' | translate }}
       </label>
       <!-- File Upload Type -->
       <div class="flex flex-col items-start gap-5 mt-2">
         <!-- UI Radio for "Send As File" -->
-        <ui-radio
+        <!-- <ui-radio
           [value]="true"
           name="uploadType"
           formControlName="selectedOption"
@@ -26,10 +21,10 @@
           <ng-container ngProjectAs="label">
             {{ 'common.notifications.email.attachment.sendFile' | translate }}
           </ng-container>
-        </ui-radio>
+        </ui-radio> -->
 
         <!-- UI Radio for "Send As URL" -->
-        <ui-radio
+        <!-- <ui-radio
           [value]="false"
           name="uploadType"
           formControlName="selectedOption"
@@ -38,10 +33,10 @@
           <ng-container ngProjectAs="label">
             {{ 'common.notifications.email.attachment.sendUrl' | translate }}
           </ng-container>
-        </ui-radio>
+        </ui-radio> -->
 
         <!-- Document properties -->
-        <form [formGroup]="propertySelectionForm">
+        <!-- <form [formGroup]="propertySelectionForm">
           <div
             *ngIf="sendAsAttachment === false"
             class="mt-2 flex flex-wrap gap-x-4 gap-y-6"
@@ -52,7 +47,6 @@
                 <span
                   class="text-red-400"
                   *ngIf="
-                    property.bodyKey === 'InformationConfidentiality' ||
                     property.bodyKey === 'DocumentType'
                   "
                 >
@@ -84,49 +78,32 @@
               </ui-select-menu>
             </div>
           </div>
-        </form>
+        </form> -->
 
         <!-- File input -->
-        <input
-          type="file"
-          id="fileUpload"
-          name="fileUpload"
-          (change)="onFileSelected($event)"
-          class="text-sm py-1 px-2 border border-gray-300 bg-white rounded-lg shadow-sm w-full"
-          multiple
-        />
+        <input type="file" id="fileUpload" name="fileUpload" (change)="onFileSelected($event)"
+          class="text-sm py-1 px-2 border border-gray-300 bg-white rounded-lg shadow-sm w-full" multiple />
       </div>
     </div>
   </div>
 
   <div>
     <!-- Provided files -->
-    <div
-      *ngIf="
+    <div *ngIf="
         this.emailService.datasetsForm?.getRawValue()?.attachments?.files
           ?.length > 0
-      "
-      class="flex flex-wrap gap-2"
-    >
-      <ui-chip
-        variant="primary"
-        *ngFor="
+      " class="flex flex-wrap gap-2">
+      <ui-chip variant="primary" *ngFor="
           let file of this.emailService.datasetsForm.getRawValue().attachments
             .files;
           let i = index
-        "
-        [removable]="true"
-        (removed)="removeFile(i)"
-      >
+        " [removable]="true" (removed)="removeFile(i)">
         {{ file.fileName }}
       </ui-chip>
     </div>
     <!-- File limitations -->
     <div>
-      <p class="text-xs italic">
-        Please note: You can upload a maximum of 5 files, with each file not
-        exceeding 7MB in size.
-      </p>
+      <p class="text-xs italic">{{ 'common.notifications.email.attachment.uploadFileNote' | translate }}</p>
     </div>
   </div>
 </div>
