+++ libs/shared/src/lib/components/email/email.component.ts
@@ -532,6 +532,7 @@ export class EmailComponent extends UnsubscribeComponent implements OnInit {
    * @param isClone Identify it is cloned or not
    */
   prepareEditData(emailData: any, isSendEmail?: boolean, isClone?: boolean) {
+    this.emailService.showFileUpload = false;
     if (isClone) {
       this.emailService.isEdit = false;
     } else {
@@ -732,7 +733,10 @@ export class EmailComponent extends UnsubscribeComponent implements OnInit {
     // Execute all queries in parallel and update metadata loading status when done
     Promise.all(promises).finally(() => {
       if (isSendEmail) {
-        this.emailService.stepperStep = 5;
+        // Show File upload when Sent email is slected
+        this.emailService.showFileUpload = this.emailService.isDirectSend;
+        // on send from EN list - navigate them to Layout screen
+        this.emailService.stepperStep = 4;
       } else {
         this.emailService.stepperStep = 0;
       }
@@ -1069,6 +1073,7 @@ export class EmailComponent extends UnsubscribeComponent implements OnInit {
    * used to create new custom template
    */
   createTemplate(): void {
+    this.emailService.showFileUpload = false;
     this.showTemplateCreationWizard = true;
     if (!this.emailService.isCustomTemplateEdit) {
       this.emailService.isNewCustomTemplate = true;
