+++ libs/shared/src/lib/components/email/steps/layout/layout.component.html
@@ -607,6 +607,10 @@
         (ngModelChange)="onEditorContentChange()"
       ></editor>
     </div>
-    <app-email-attachment></app-email-attachment>
+
+    <!-- On email notification creation and edit - Hide File upload in Step 5 -->
+    <ng-container *ngIf="emailService.showFileUpload">
+      <app-email-attachment></app-email-attachment>
+    </ng-container>
   </div>
 </form>
