+++ libs/shared/src/lib/components/email/steps/layout/layout.component.ts
@@ -11,6 +11,7 @@ import { Subscription } from 'rxjs';
 import { UnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
 import convertToMinutes from '../../../../utils/convert-to-minutes';
 import { COMMA, ENTER, SPACE, TAB } from '@angular/cdk/keycodes';
+import { cloneDeep } from 'lodash';
 /**
  * Email layout page component.
  */
@@ -166,6 +167,17 @@ export class LayoutComponent
       const layoutName = this.emailService?.emailLayout?.name || '';
       this.emailService.emailLayout = {};
       this.emailService.emailLayout.name = layoutName;
+
+      //Bind DL data
+      if (this.emailService.quickEmailDLQuery?.length === 0) {
+        const toData = this.emailService.emailDistributionList?.to;
+        const ccData = this.emailService.emailDistributionList?.cc;
+        const bccData = this.emailService.emailDistributionList?.bcc;
+        this.emailService.quickEmailDLQuery = { to: [], cc: [], bcc: [] };
+        this.emailService.quickEmailDLQuery.to = cloneDeep(toData);
+        this.emailService.quickEmailDLQuery.cc = cloneDeep(ccData);
+        this.emailService.quickEmailDLQuery.bcc = cloneDeep(bccData);
+      }
     } else {
       this.emailService.createPreviewData();
     }
@@ -254,22 +266,66 @@ export class LayoutComponent
           this.emailService.allLayoutdata.footerLogo;
       }
     }
-    if (this.emailService.emailDistributionList.to) {
-      this.layoutForm
-        .get('to')
-        ?.setValue(this.emailService.emailDistributionList.to);
-    }
-    if (this.emailService.emailDistributionList.cc) {
-      this.layoutForm
-        .get('cc')
-        ?.setValue(this.emailService.emailDistributionList.cc);
-    }
-    if (this.emailService.emailDistributionList.bcc) {
-      this.layoutForm
-        .get('bcc')
-        ?.setValue(this.emailService.emailDistributionList.bcc);
+    if (this.emailService.isQuickAction) {
+      this.populateDLForm();
     }
     this.getBlockData();
+  }
+
+  /**
+   * Populates dataset Form using custom template DL object
+   */
+  populateDLForm() {
+    if (this.emailService.isQuickAction) {
+      const { to, cc, bcc } = this.emailService.emailDistributionList; //this.emailService.customLayoutDL;
+
+      const uniqueTo: any = [...new Set(to?.inputEmails ?? to)];
+      const uniqueCc: any = [...new Set(cc?.inputEmails ?? cc)];
+      const uniqueBcc: any = [...new Set(bcc?.inputEmails ?? bcc)];
+
+      this.emailService.emailDistributionList.to = uniqueTo;
+      this.emailService.emailDistributionList.cc = uniqueCc;
+      this.emailService.emailDistributionList.bcc = uniqueBcc;
+
+      this.emailService.populateEmails(
+        this.emailService.emailDistributionList.to,
+        this.emailService?.datasetsForm
+          ?.get('emailDistributionList')
+          ?.get('to')
+          ?.get('inputEmails') as FormArray
+      );
+
+      this.emailService.populateEmails(
+        this.emailService.emailDistributionList.cc,
+        this.emailService?.datasetsForm
+          ?.get('emailDistributionList')
+          ?.get('cc')
+          ?.get('inputEmails') as FormArray
+      );
+
+      this.emailService.populateEmails(
+        this.emailService.emailDistributionList.bcc,
+        this.emailService?.datasetsForm
+          ?.get('emailDistributionList')
+          ?.get('bcc')
+          ?.get('inputEmails') as FormArray
+      );
+      if (this.emailService.emailDistributionList.to) {
+        this.layoutForm
+          .get('to')
+          ?.setValue(this.emailService.emailDistributionList.to);
+      }
+      if (this.emailService.emailDistributionList.cc) {
+        this.layoutForm
+          .get('cc')
+          ?.setValue(this.emailService.emailDistributionList.cc);
+      }
+      if (this.emailService.emailDistributionList.bcc) {
+        this.layoutForm
+          .get('bcc')
+          ?.setValue(this.emailService.emailDistributionList.bcc);
+      }
+    }
   }
 
   /**
