+++ libs/shared/src/lib/components/templates/components/preview-template-modal/preview-template-modal.component.html
@@ -43,7 +43,7 @@
         *ngIf="currentStep !== 1"
         type="submit"
         category="secondary"
-        [disabled]="disableActionButton"
+        [disabled]="disableActionButton || emailService.disableNextActionBtn"
         (click)="next()"
       >
         {{ 'common.next' | translate }}
