+++ libs/shared/src/lib/components/widgets/grid-settings/button-config/button-config.component.ts
@@ -421,6 +421,7 @@ export class ButtonConfigComponent
 
   /** Opens modal for adding a new email template */
   public async addEmailTemplate() {
+    this.emailService.showFileUpload = false;
     this.emailService.resetAllLayoutData();
     const { TemplateModalComponent } = await import(
       '../../../templates/components/template-modal/template-modal.component'
